requirements.txt
atlassian-python-api>=3.32.0
click>=8.1.3
tqdm>=4.64.1
python-dotenv>=1.0.0
